+++ src/capturer/custom.rs
@@ -2,12 +2,11 @@ use super::{capture, model::Capturer};
 use crate::{Error, Monitor, OutDuplDescExt, Result};
 use windows::Win32::Graphics::{
   Direct3D11::{ID3D11Texture2D, D3D11_TEXTURE2D_DESC},
-  Dxgi::{DXGI_OUTDUPL_FRAME_INFO, DXGI_OUTDUPL_POINTER_SHAPE_INFO, DXGI_OUTPUT_DESC},
+  Dxgi::{DXGI_OUTDUPL_FRAME_INFO, DXGI_OUTDUPL_POINTER_SHAPE_INFO},
 };
 
 /// Capture screen to a chunk of memory.
 pub struct CustomCapturer<'a> {
-  // TODO: use a raw pointer?
   buffer: &'a mut [u8],
   monitor: Monitor,
   texture: ID3D11Texture2D,
@@ -41,12 +40,8 @@ impl<'a> CustomCapturer<'a> {
 }
 
 impl Capturer for CustomCapturer<'_> {
-  fn dxgi_output_desc(&self) -> Result<DXGI_OUTPUT_DESC> {
-    self.monitor.dxgi_output_desc()
-  }
-
-  fn dxgi_outdupl_desc(&self) -> windows::Win32::Graphics::Dxgi::DXGI_OUTDUPL_DESC {
-    self.monitor.dxgi_outdupl_desc()
+  fn monitor(&self) -> &Monitor {
+    &self.monitor
   }
 
   fn buffer(&self) -> &[u8] {
@@ -58,7 +53,7 @@ impl Capturer for CustomCapturer<'_> {
   }
 
   fn check_buffer(&self) -> Result<()> {
-    if self.buffer.len() < self.dxgi_outdupl_desc().calc_buffer_size() {
+    if self.buffer.len() < self.monitor.dxgi_outdupl_desc().calc_buffer_size() {
       Err(Error::InvalidBufferLength)
     } else {
       Ok(())